Auto stash before merge of "dev/IFX007T_driver_PM" and "origin/dev/IFX007T_driver_PM"
